Make sure vms are shut off before backup them
We need to make sure that the vms are really down
before tar and backup them.
This cause some failure with the undercloud, which is
pretty slow to go down

In order to do that we also need to run the playbook
with stack user instead of root

Need also to make sure that the vms are not set as
autostart.

Change-Id: Ia6626eb373f98af7ab12fdee476487d0d9e5cadb